data_mainetance/requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
fsspec==0.7.3
gcsfs==0.6.1
google-api-core==1.17.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
lxml==4.5.0
numpy==1.18.3
oauthlib==3.1.0
pandas==1.0.3
protobuf==3.11.3
psycopg2-binary==2.8.5
pyarrow==0.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
urllib3==1.25.9